refactor: 피드백 반영
피드백 사항에 맞춰 로직을 변경했습니다.
