+++ BE/src/main/java/team05/codesquad/issuetracker/service/LabelService.java
@@ -5,13 +5,14 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import team05.codesquad.issuetracker.controller.labeldto.LabelRequest;
-import team05.codesquad.issuetracker.controller.labeldto.LabelResponse;
-import team05.codesquad.issuetracker.controller.labeldto.LabelsResponse;
+import team05.codesquad.issuetracker.controller.labeldto.request.LabelRequest;
+import team05.codesquad.issuetracker.controller.labeldto.response.LabelResponse;
+import team05.codesquad.issuetracker.controller.labeldto.response.LabelsResponse;
 import team05.codesquad.issuetracker.domain.label.Label;
 import team05.codesquad.issuetracker.repository.LabelRepository;
 
-import javax.lang.model.type.ErrorType;
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 @Transactional
@@ -22,17 +23,19 @@ public class LabelService {
 
     public LabelResponse save(LabelRequest request) {
         Label label = request.toEntity();
-        Label saveLabel = labelRepository.save(label);
-        return LabelResponse.from(saveLabel);
+        return LabelResponse.from(labelRepository.save(label));
     }
 
-    public Iterable<Label> findAll() {
-        return labelRepository.findAll();
+    public LabelsResponse findAll() {
+        return LabelsResponse.from(labelRepository.findAll());
     }
 
+
+    // 업데이트하고, 업데이트 한 값을 반환해주는 역할 -> editLabel
     public LabelResponse edit(Long id, LabelRequest request) {
         Label targetLabel = labelRepository.findById(id).orElseThrow(() -> new NotFoundException(HttpStatus.NOT_FOUND + "정보를 찾을 수 없습니다"));
         targetLabel.editLabel(request.getTitle(), request.getDescription(), request.getBackgroundColor(), request.getFontColor());
+        labelRepository.update(targetLabel.getId(), targetLabel.getTitle(), targetLabel.getDescription(), targetLabel.getBackgroundColor(), targetLabel.getFontColor());
         return LabelResponse.from(targetLabel);
     }
